refactor(plugin_isolation): Split plugin isolation module
Refactored plugin_isolation module to follow single responsibility principle:

**Module Structure (560 lines → 5 files):**
- types.rs (147 lines): PluginState, PluginMetadata, ResourceLimits, PluginInstance, PluginMetrics
- config.rs (109 lines): PluginManagerConfig, SecurityPolicy, MonitoringConfig, IsolationConfig
- manager.rs (373 lines): IsolatedPluginManager implementation
- health.rs (21 lines): PluginManagerHealth
- mod.rs (43 lines): Re-exports only

**Benefits:**
- Clear separation of types, configuration, and implementation
- Configuration centralized in dedicated module
- Health information isolated for easy monitoring
- Manager logic focused on orchestration

**Testing:**
- All 377 tests passing
- No clippy warnings
- Formatting check passed
- Public API unchanged

Closes #164
